services/gateway-service/requirements.txt
# FastAPI Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
psycopg2-binary==2.9.9

# HTTP Client (Proxy)
httpx==0.27.2

# Data Validation
pydantic==2.9.2

# Environment Variables
python-dotenv==1.0.1
